Rename examples scroll area component to kebab-case
