--- packages/examples/src/components/ui/ScrollArea/index.ts
@@ -1,2 +1,0 @@
-export * from './ScrollArea';
-export * from './ScrollBar';
