+++ packages/examples/src/components/ui/scroll-area/index.ts
@@ -1,0 +1,2 @@
+export * from './scroll-area';
+export * from './scrollbar';
