+++ packages/examples/src/components/ui/scroll-area/scrollbar.tsx
